Fix sticky table headers in Safari
In Safari, issues exist with table headers. Safari does not appear to honor the “width” CSS directive in this case, preferring to automatically adjust to what it thinks is the best width (which is wrong). Setting min-width and max-width resolves this.

Adding the <br> tag to the template fixes an issue in which comments appear on the same line as column headers when the screen is scrolled.

Moved is_firefox to higher scope to avoid its calculation each time in the loop.